refactor(agent): extract leaf modules (cmdline/vsock/probe/btrfs)
Pull four leaf utility groups out of linux/mod.rs into siblings:

- cmdline.rs: cmdline_value + the docker/kubernetes vsock-port and
  data-device readers (HVC fast-path device).
- vsock.rs: bind_vsock_listener_with_retry and is_peer_closed_error.
- probe.rs: TCP / Unix-socket / Docker-API probes.
- btrfs.rs: data-volume detection, format, and bind-mount setup,
  including the BTRFS_IOC_SUBVOL_CREATE ioctl and NOCOW pass.

Each fn keeps `pub(super)` visibility — they are only used inside
linux/. The non-Linux stubs of disable_cow_on_metadata_dirs and
btrfs_create_subvolume are dropped: btrfs.rs lives under the existing
`#[cfg(target_os = "linux")]`-gated linux module so the stubs were
already dead code.

linux/mod.rs drops the now-unused use lines (cmdline / devices / env
constants, std::io seek/read traits, and a few tokio_vsock items)
and gains four `mod` + `use` lines for the new siblings.

Pure mechanical move. cargo check/clippy emit the same warning set
as before the commit.
